Tweaks based upon feedback (javadoc, naming, etc)
